requirements.txt
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.2
python-dotenv==1.0.1
scikit-learn==1.2.1
streamlit